security: fix buffer overflows, integer overflow, heap underflow, PFA frame 0, add -Werror
- kernel.c: bounds check on echo redirection, history overflow protection,
  integer overflow guards in hex/sleep, fix sign-compare warnings
- heap.c: guard prev_size read against underflow below HEAP_START
- pfa.c: remove separate mark_frame(0), now covered by loop from 0x00000000
- screen.c: capture buffer bounds protection (set_capture/get_capture)
- Makefile: add -Werror to CFLAGS
- docs: update README and CODEBASE_INDEX with security hardening notes

+++ kernel/kernel.c
@@ -238,7 +238,7 @@ static void execute_cmd(const char *cmd, char *arg)
         unsigned int n = 0;
         int k = 0;
         while (arg[k]) {
-            if (k >= 8 || n > UINT_MAX / 10 || (n == UINT_MAX / 10 && (arg[k] - '0') > UINT_MAX % 10)) {
+            if (k >= 8 || n > UINT_MAX / 10 || (n == UINT_MAX / 10 && (unsigned int)(arg[k] - '0') > UINT_MAX % 10)) {
                 print_string("Invalid number\n");
                 return;
             }
@@ -253,7 +253,7 @@ static void execute_cmd(const char *cmd, char *arg)
         unsigned int n = 0;
         int k = 0;
         while (arg[k]) {
-            if (k >= 7 || n > UINT_MAX / 10 || (n == UINT_MAX / 10 && (arg[k] - '0') > UINT_MAX % 10)) {
+            if (k >= 7 || n > UINT_MAX / 10 || (n == UINT_MAX / 10 && (unsigned int)(arg[k] - '0') > UINT_MAX % 10)) {
                 print_string("Invalid number\n");
                 return;
             }
